scripts/requirements.txt
# rewrite_note_sources.py
PyYAML>=6.0.1
